fix(backend): enforce materialization contracts
