+++ src/backend/tests/community/architecture/test_service_api_conformance.py
@@ -94,6 +94,9 @@ from agentclaw.community.api.local_skill_delete_service import (
 from agentclaw.community.api.skill_set_management_service import (
     SkillSetManagementServiceProtocol,
 )
+from agentclaw.community.api.skill_version_materializer import (
+    SkillVersionMaterializerProtocol,
+)
 from agentclaw.community.api.market_favorite_service import (
     MarketFavoriteServiceProtocol,
 )
@@ -173,6 +176,9 @@ from agentclaw.community.core.skill_center.services.repository_catalog_service i
 from agentclaw.community.core.skill_center.services.skill_set_management_service import (
     SkillSetManagementService,
 )
+from agentclaw.community.core.skill_center.services.skill_version_materializer import (
+    SkillVersionMaterializer,
+)
 from agentclaw.community.core.skill_center.services.space_skill_grant_service import (
     SpaceSkillGrantService,
 )
@@ -220,6 +226,7 @@ _PAIRS = [
     (LocalSkillDeleteServiceProtocol, LocalSkillDeleteService),
     (RepositoryCatalogServiceProtocol, RepositoryCatalogService),
     (SkillSetManagementServiceProtocol, SkillSetManagementService),
+    (SkillVersionMaterializerProtocol, SkillVersionMaterializer),
     (SpaceServiceProtocol, SpaceService),
     (SpaceAccessServiceProtocol, SpaceAccessService),
     (SpaceMemberServiceProtocol, SpaceMemberService),
